+++ db/migrate/20180124195233_add_forum_auto_subscribed_to_users.rb
@@ -1,0 +1,6 @@
+class AddForumAutoSubscribedToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :forum_auto_subscribed, :boolean, null: false, default: false
+    add_index :users, :forum_auto_subscribed
+  end
+end